+++ CHANGELOG.md
@@ -4,9 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.1.1] / 2022-04-22
+## [0.1.1] / 2022-04-25
+### Features
+- Remove RibbonTab name from Revit Dictionary - `GetRibbonTabsDictionary`
 ### Added
-- Add `Autodesk.Windows.ComponentManager.IsApplicationFrameEnabled` to disable buttons on modeless
+- Add `Autodesk.Windows.ComponentManager.IsApplicationFrameEnabled` to disable buttons on modeless Example
 
 ## [0.1.0] / 2022-04-04
 ### Added